% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lesson_path.R
\name{get_lesson_path}
\alias{get_lesson_path}
\title{Get lesson path}
\usage{
get_lesson_path()
}
\value{
Path of lesson.yaml file for current lesson (character)
}
\description{
Gets path of lesson.yaml file for current lesson
}
\author{
fnaufel
}
