% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scripts_path.R
\name{get_scripts_path}
\alias{get_scripts_path}
\title{Build path name of scripts directory}
\usage{
get_scripts_path()
}
\value{
Path name of scripts directory of current lesson (character).
}
\description{
Builds path name of scripts directory of current lesson.
}
\author{
fnaufel
}
